Separate out queries and seeding functions and refactor to manage async code
