+++ db/seed.js
@@ -1,55 +1,34 @@
+const db = require('./index.js');
+const faker = require('faker');
 const mysql = require('mysql');
-const config = require('./config.js');
 const connection = mysql.createConnection(config);
 
-const insertUser  = (username) => {
-  return new Promise((resolve, reject) => {
-    let sql = 'INSERT INTO users (username) VALUES (?)'
-    connection.query(sql, username, (err, result) => {
-      if (err) { reject(err) }
-      resolve(result)
-    })
-  })
+
+const seedUsers = async () => {
+  for (let i = 0; i < 100; i++) {
+    let user = {username: faker.internet.userName()}
+    await db.insertUser(user);
+  }
 }
 
-const insertReview = (date, review, rating, book_id, user_id) => {
-  return new Promise((resolve, reject) => {
-    let sql = 'insert into reviews (date, review, rating, book_id, user_id) values (?, ?, ?, ?, ?)';
-    let params = [date, review, rating, book_id, user_id];
-    connection.query(sql, params, (err, result) => {
-      if (err) { reject(err) }
-      resolve(result)
-    })
-  })
+const seedReviews = async () => {
+  for (let j = 0; j < 100; j++) {
+    let review = {};
+    review.date = faker.date.past();
+    review.review = faker.lorem.paragraph();
+    review.rating = ( Math.floor(Math.random() * 5) + 1 );
+    review.book_id = ( Math.floor(Math.random() * 100) + 1 );
+    review.user_id = ( Math.floor(Math.random() * 100) + 1 );
+
+    await db.insertReview(review);
+  }
+
+const seedAllData = async () => {
+  await seedUsers();
+  await seedReviews();
 }
 
-const getReviews = (id) => {
-  return new Promise((resolve, reject) => {
-    const sql = 'select * from reviews where id = ?'
-    connection.query(sql, id, (err, result) => {
-      if (err) { reject (err) }
-      resolve(result)
-    })
+seedAllData().then(() => {
+  connection.end();
   })
-}
-
-
-
-const getRatedReviews = (id, rating) => {
-  return new Promise((resolve, reject) => {
-    const sql = "select * from reviews where (id = ?, rating = ?)"
-    let params = [id, rating];
-    connection.query(sql, params, (err, result) => {
-      if (err) { reject (err) }
-      resolve(result)
-    });
-  })
-}
-
-
-module.exports = {
-  insertUser,
-  insertReview,
-  getReviews,
-  getRatedReviews
 }
